Moved oauth component. Created a separate entry point for component definitions.
