+++ libs/platform/experience/layout/src/stories/static/examples/product-images.stories.ts
@@ -1,7 +1,7 @@
 import { Story } from '@storybook/web-components';
 
 import { html, TemplateResult } from 'lit';
-import { storybookPrefix } from '../../../../../constants';
+import { storybookPrefix } from '../../../../../.constants';
 import { generateLayoutItems } from '../util';
 import { pageStyles } from './util';
