+++ libs/platform/experience/layout/src/stories/static/list.stories.ts
@@ -1,7 +1,7 @@
 import { CompositionLayout } from '@spryker-oryx/experience';
 import { Story } from '@storybook/web-components';
 import { html, TemplateResult } from 'lit';
-import { storybookPrefix } from '../../../../constants';
+import { storybookPrefix } from '../../../../.constants';
 import { layoutStaticStyles } from './styles';
 import { generateLayoutItems, generateNestedLayout } from './util';
